I've updated a few things to increase the scores in our Lighthouse report. @ananavati @donsalari

- All images should have alt tags
- Our templates need viewport tags
- Our main JS bundle tag should be at the bottom of the body

Doing these few things brought down our first meaningful paint time from 3500ms to 720ms and our time to interactive from 4500ms to 960ms. Our last goal for perf should be to further increase our perceptual speed index score. 

@aweary I'll also put in a PR to update the boilerplate with these perf goals in mind.

![screen shot 2016-11-17 at 3 03 46 pm](https://cloud.githubusercontent.com/assets/2831294/20411625/206517a6-acd7-11e6-800f-b163aa1aecdc.png)