@ananavati 2016-11-17
interesting perf improvements, wonder how these numbers match up to https://www.webpagetest.org/ once we deploy this app.